+++ InnerEye/Azure/azure_runner.py
@@ -18,7 +18,7 @@ from InnerEye.Azure.secrets_handling import read_all_settings
 from InnerEye.Common.generic_parsing import GenericConfig
 from InnerEye.ML.common import ModelExecutionMode
 from InnerEye.ML.utils.config_loader import ModelConfigLoader
-from health.azure.datasets import DatasetConfig
+from health_azure import DatasetConfig
 
 SLEEP_TIME_SECONDS = 30