+++ Tests/Common/test_commandline_parsing.py
@@ -15,7 +15,7 @@ from InnerEye.ML.config import PhotometricNormalizationMethod, SegmentationModel
 from InnerEye.ML.deep_learning_config import DeepLearningConfig
 from InnerEye.ML.runner import Runner
 from Tests.ML.configs.DummyModel import DummyModel
-from health.azure.himl import AzureRunInfo
+from health_azure import AzureRunInfo
 
 
 @pytest.mark.parametrize("is_container", [True, False])
@@ -65,7 +65,7 @@ def test_create_ml_runner_args(is_container: bool,
                     azure_run_info = AzureRunInfo(input_datasets=[None],
                                                   output_datasets=[None],
                                                   run=None,
-                                                  is_running_in_azure=False,
+                                                  is_running_in_azure_ml=False,
                                                   output_folder=Path.cwd(),
                                                   logs_folder=Path.cwd())
                     runner.run_in_situ(azure_run_info)